third_party/web_platform_tests/workers/interfaces/WorkerUtils/WindowTimers/003.html

<!-- Pulled from chromium at commit: https://source.chromium.org/chromium/chromium/src/+/main:third_party/blink/web_tests/external/wpt/workers/interfaces/WorkerUtils/WindowTimers/003.html;drc=e5b8fefe2aae5674dc00817d444a50f67bdb6ea0 -->
<!doctype html>
<title>setInterval</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<div id=log></div>
<script>
  async_test(function () {
    var worker = new Worker('003.js');
    worker.onmessage = this.step_func(function (e) {
      assert_equals(e.data, '1');
      this.done();
    });
  });
</script>
